pyproject.toml
[project]
name = "scrainshots"
version = "0.1.0"
description = "AI-powered screenshot analysis and management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "pillow>=10.2.0",
    "python-dateutil>=2.8.2",
    "tqdm>=4.66.1",
    "chromadb>=0.4.22",
    "langchain>=0.1.5",
    "langchain-community>=0.0.16",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]

[project.optional-dependencies]
local = [
    "mlx>=0.0.10",
    "mlx-lm>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["*.py"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ["py38"]